Compute version in tools.py to make it available to other commands
